refactor: migrate VersionInfo to builder binary
Move VersionInfo struct and register_version_metrics from op-rbuilder
crate to the builder binary. This decouples version metrics registration
from the crate, allowing binaries to control their own version reporting.

- Add bin/builder/src/version.rs with VersionInfo and metrics registration
- Update launcher to accept on_node_started callback
- Move SHORT_VERSION/LONG_VERSION constants to args/mod.rs
- Simplify op-rbuilder build.rs to only generate CLI version env vars
- Remove VersionInfo from metrics.rs

Closes #396
