+++ bin/builder/src/main.rs
@@ -3,6 +3,8 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+pub mod version;
+
 #[global_allocator]
 static ALLOC: reth_cli_util::allocator::Allocator = reth_cli_util::allocator::new_allocator();
 
@@ -10,7 +12,9 @@ fn main() {
     base_cli_utils::Backtracing::enable();
     base_cli_utils::SigsegvHandler::install();
 
-    if let Err(err) = op_rbuilder::launcher::launch() {
+    if let Err(err) = op_rbuilder::launcher::launch(|| {
+        version::VERSION.register_version_metrics();
+    }) {
         eprintln!("Error: {err:?}");
         std::process::exit(1);
     }
