+++ bin/builder/src/version.rs
@@ -1,0 +1,104 @@
+//! Version information for the op-rbuilder binary.
+//!
+//! Contains [`VersionInfo`] and version-related constants.
+//!
+//! Derived from [`reth-node-core`'s type][reth-version-info]
+//!
+//! [reth-version-info]: https://github.com/paradigmxyz/reth/blob/805fb1012cd1601c3b4fe9e8ca2d97c96f61355b/crates/node/metrics/src/version.rs#L6
+
+use metrics::gauge;
+
+/// Cargo package version.
+pub const CARGO_PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
+
+/// Build timestamp from vergen.
+pub const VERGEN_BUILD_TIMESTAMP: &str = env!("VERGEN_BUILD_TIMESTAMP");
+
+/// Cargo features from vergen.
+pub const VERGEN_CARGO_FEATURES: &str = env!("VERGEN_CARGO_FEATURES");
+
+/// Git SHA (short) from vergen.
+pub const VERGEN_GIT_SHA: &str = env!("VERGEN_GIT_SHA_SHORT");
+
+/// Cargo target triple from vergen.
+pub const VERGEN_CARGO_TARGET_TRIPLE: &str = env!("VERGEN_CARGO_TARGET_TRIPLE");
+
+/// Build profile name.
+pub const BUILD_PROFILE_NAME: &str = env!("OP_RBUILDER_BUILD_PROFILE");
+
+/// The latest commit author.
+pub const VERGEN_GIT_AUTHOR: &str = env!("VERGEN_GIT_COMMIT_AUTHOR");
+
+/// The latest commit message.
+pub const VERGEN_GIT_COMMIT_MESSAGE: &str = env!("VERGEN_GIT_COMMIT_MESSAGE");
+
+/// Short version string.
+pub const SHORT_VERSION: &str = env!("OP_RBUILDER_SHORT_VERSION");
+
+/// Long version string with additional build info.
+pub const LONG_VERSION: &str = concat!(
+    env!("OP_RBUILDER_LONG_VERSION_0"),
+    "\n",
+    env!("OP_RBUILDER_LONG_VERSION_1"),
+    "\n",
+    env!("OP_RBUILDER_LONG_VERSION_2"),
+    "\n",
+    env!("OP_RBUILDER_LONG_VERSION_3"),
+    "\n",
+    env!("OP_RBUILDER_LONG_VERSION_4"),
+    "\n",
+    env!("OP_RBUILDER_LONG_VERSION_5"),
+);
+
+/// Contains version information for the application and allows for exposing the contained
+/// information as a prometheus metric.
+#[derive(Debug, Clone)]
+pub struct VersionInfo {
+    /// The version of the application.
+    pub version: &'static str,
+    /// The build timestamp of the application.
+    pub build_timestamp: &'static str,
+    /// The cargo features enabled for the build.
+    pub cargo_features: &'static str,
+    /// The Git SHA of the build.
+    pub git_sha: &'static str,
+    /// The target triple for the build.
+    pub target_triple: &'static str,
+    /// The build profile (e.g., debug or release).
+    pub build_profile: &'static str,
+    /// The author of the latest commit.
+    pub commit_author: &'static str,
+    /// The message of the latest commit.
+    pub commit_message: &'static str,
+}
+
+/// The version information for op-rbuilder.
+pub const VERSION: VersionInfo = VersionInfo {
+    version: CARGO_PKG_VERSION,
+    build_timestamp: VERGEN_BUILD_TIMESTAMP,
+    cargo_features: VERGEN_CARGO_FEATURES,
+    git_sha: VERGEN_GIT_SHA,
+    target_triple: VERGEN_CARGO_TARGET_TRIPLE,
+    build_profile: BUILD_PROFILE_NAME,
+    commit_author: VERGEN_GIT_AUTHOR,
+    commit_message: VERGEN_GIT_COMMIT_MESSAGE,
+};
+
+impl VersionInfo {
+    /// This exposes op-rbuilder's version information over prometheus.
+    pub fn register_version_metrics(&self) {
+        let labels: [(&str, &str); 8] = [
+            ("version", self.version),
+            ("build_timestamp", self.build_timestamp),
+            ("cargo_features", self.cargo_features),
+            ("git_sha", self.git_sha),
+            ("target_triple", self.target_triple),
+            ("build_profile", self.build_profile),
+            ("commit_author", self.commit_author),
+            ("commit_message", self.commit_message),
+        ];
+
+        let gauge = gauge!("builder_info", &labels);
+        gauge.set(1);
+    }
+}
